feat(items): non-billable items — free accompaniments & supplies reach POS uncharged
- Item.non_billable (Ent bool, default false): never charged at the point of
  sale even when a selling price exists (free accompaniments like ugali,
  consumable supplies like tissue/packaging); stock still deducts on sale.
- ItemDTO carries non_billable (pointer — partial updates never clobber);
  create/update set it only when the client sends it.
- ListItems ?include_non_billable=1 (context flag, mirrors include=variants)
  ORs non-billable items into the type filter so the POS catalog proxy can
  surface supplies whose item type is outside the use-case's sellable types.
- Atlas migration (idempotent ADD COLUMN IF NOT EXISTS — also captures
  previously-unversioned e-commerce/PIN/supplier column drift) + backfill
  marking ugali items non-billable.
- Fix pre-existing stock-availability test asserting int against float64
  quantities.

+++ internal/http/handlers/inventory_test.go
@@ -291,9 +291,10 @@ func TestGetStockAvailability_Success(t *testing.T) {
 	var avail items.StockAvailability
 	require.NoError(t, json.NewDecoder(w.Body).Decode(&avail))
 	assert.Equal(t, "LATTE-001", avail.SKU)
-	assert.Equal(t, 50, avail.OnHand)
-	assert.Equal(t, 45, avail.Available)
-	assert.Equal(t, 5, avail.Reserved)
+	// Stock quantities are float64 (fractional units like 4.5 L are valid).
+	assert.Equal(t, float64(50), avail.OnHand)
+	assert.Equal(t, float64(45), avail.Available)
+	assert.Equal(t, float64(5), avail.Reserved)
 }
 
 func TestGetStockAvailability_NotFound(t *testing.T) {
